# apps/web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccessAction": {
        "enum": [
          "view",
          "create",
          "edit",
          "export",
          "print"
        ],
        "title": "AccessAction",
        "type": "string"
      },
      "AccessDecision": {
        "enum": [
          "allow",
          "deny"
        ],
        "title": "AccessDecision",
        "type": "string"
      },
      "AccessLogEntry": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/AccessAction"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "because": {
            "$ref": "#/components/schemas/Because"
          },
          "break_glass": {
            "title": "Break Glass",
            "type": "boolean"
          },
          "decision": {
            "$ref": "#/components/schemas/AccessDecision"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "query_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "resource": {
            "$ref": "#/components/schemas/ResourceType"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "user_name": {
            "title": "User Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "at",
          "user_name",
          "role",
          "resource",
          "action",
          "decision",
          "break_glass",
          "because",
          "query_status"
        ],
        "title": "AccessLogEntry",
        "type": "object"
      },
      "AccessLogPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccessLogEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "AccessLogPage",
        "type": "object"
      },
...
      "AllergyOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "reaction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reaction"
          },
          "severity": {
            "$ref": "#/components/schemas/AllergySeverity"
          },
          "substance": {
            "title": "Substance",
            "type": "string"
          }
        },
        "required": [
          "id",
          "substance",
          "reaction",
          "severity",
          "is_active"
        ],
        "title": "AllergyOut",
        "type": "object"
      },
      "AllergySeverity": {
        "enum": [
          "mild",
          "moderate",
          "severe"
        ],
        "title": "AllergySeverity",
        "type": "string"
      },
...
      "AppointmentKind": {
        "enum": [
          "new",
          "followup",
          "walkin"
        ],
        "title": "AppointmentKind",
        "type": "string"
      },
...
      "AppointmentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "doctor_user_id": {
            "format": "uuid",
            "title": "Doctor User Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AppointmentKind"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "slot_end": {
            "format": "date-time",
            "title": "Slot End",
            "type": "string"
          },
          "slot_start": {
            "format": "date-time",
            "title": "Slot Start",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/AppointmentSource"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          }
        },
        "required": [
          "id",
          "patient_id",
          "doctor_user_id",
          "slot_start",
          "slot_end",
          "kind",
          "status",
          "source",
          "created_at"
        ],
        "title": "AppointmentOut",
        "type": "object"
      },
...
      "Because": {
        "description": "Why an access was expected, with the details the patient needs to recognise it.",
        "properties": {
          "from_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Name"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "slot_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Start"
          },
          "template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template"
          },
          "tests": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tests"
          },
          "token_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token No"
          }
        },
        "required": [
          "template"
        ],
        "title": "Because",
        "type": "object"
      },
...
      "ConditionOut": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "onset": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onset"
          },
          "status": {
            "$ref": "#/components/schemas/ConditionStatus"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "text",
          "onset",
          "status"
        ],
        "title": "ConditionOut",
        "type": "object"
      },
...
      "ConsentChannel": {
        "enum": [
          "reception",
          "patient_portal"
        ],
        "title": "ConsentChannel",
        "type": "string"
      },
      "ConsentEventKind": {
        "enum": [
          "granted",
          "withdrawn"
        ],
        "title": "ConsentEventKind",
        "type": "string"
      },
      "ConsentEventOut": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "by_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "By User Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ConsentEventKind"
          }
        },
        "required": [
          "id",
          "kind",
          "at",
          "by_user_id"
        ],
        "title": "ConsentEventOut",
        "type": "object"
      },
      "ConsentOut": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/ConsentChannel"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ConsentEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "granted_at": {
            "format": "date-time",
            "title": "Granted At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notice_id": {
            "format": "uuid",
            "title": "Notice Id",
            "type": "string"
          },
          "notice_version": {
            "title": "Notice Version",
            "type": "integer"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "withdrawn_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn At"
          }
        },
        "required": [
          "id",
          "patient_id",
          "notice_id",
          "notice_version",
          "granted_at",
          "channel",
          "withdrawn_at",
          "events"
        ],
        "title": "ConsentOut",
        "type": "object"
      },
      "ConsentStatus": {
        "properties": {
          "granted_at": {
            "format": "date-time",
            "title": "Granted At",
            "type": "string"
          },
          "notice_version": {
            "title": "Notice Version",
            "type": "integer"
          },
          "withdrawn_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn At"
          }
        },
        "required": [
          "notice_version",
          "granted_at",
          "withdrawn_at"
        ],
        "title": "ConsentStatus",
        "type": "object"
      },
...
      "DocumentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "sha256",
          "created_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "total_paise": {
            "title": "Total Paise",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "total_paise",
          "created_at"
        ],
        "title": "InvoiceSummary",
        "type": "object"
      },
...
      "PatientChart": {
        "description": "What a patient sees of their own record: signed notes and prescriptions, released results.",
        "properties": {
          "allergies": {
            "items": {
              "$ref": "#/components/schemas/AllergyOut"
            },
            "title": "Allergies",
            "type": "array"
          },
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/AppointmentOut"
            },
            "title": "Appointments",
            "type": "array"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ConditionOut"
            },
            "title": "Conditions",
            "type": "array"
          },
          "consent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "title": "Documents",
            "type": "array"
          },
          "invoices": {
            "items": {
              "$ref": "#/components/schemas/InvoiceSummary"
            },
            "title": "Invoices",
            "type": "array"
          },
          "lab_orders": {
            "items": {
              "$ref": "#/components/schemas/LabOrderOut"
            },
            "title": "Lab Orders",
            "type": "array"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteOut"
            },
            "title": "Notes",
            "type": "array"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientOut"
          },
          "prescriptions": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionOut"
            },
            "title": "Prescriptions",
            "type": "array"
          },
          "view": {
            "const": "patient",
            "default": "patient",
            "title": "View",
            "type": "string"
          },
          "vitals": {
            "items": {
              "$ref": "#/components/schemas/VitalOut"
            },
            "title": "Vitals",
            "type": "array"
          }
        },
        "required": [
          "patient",
          "appointments",
          "vitals",
          "allergies",
          "conditions",
          "notes",
          "prescriptions",
          "lab_orders",
          "documents",
          "invoices",
          "consent"
        ],
        "title": "PatientChart",
        "type": "object"
      },
...
      "PatientExport": {
        "properties": {
          "access_log": {
            "items": {
              "$ref": "#/components/schemas/AccessLogEntry"
            },
            "title": "Access Log",
            "type": "array"
          },
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentOut"
            },
            "title": "Consents",
            "type": "array"
          },
          "exported_at": {
            "format": "date-time",
            "title": "Exported At",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "record": {
            "$ref": "#/components/schemas/PatientChart"
          }
        },
        "required": [
          "exported_at",
          "record",
          "consents",
          "access_log",
          "note"
        ],
        "title": "PatientExport",
        "type": "object"
      },
      "PatientOut": {
        "properties": {
          "abha_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abha Number"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dob": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "emergency_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mrn": {
            "title": "Mrn",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          }
        },
        "required": [
          "id",
          "mrn",
          "name",
          "dob",
          "sex",
          "phone",
          "address",
          "abha_number",
          "emergency_contact",
          "created_at"
        ],
        "title": "PatientOut",
        "type": "object"
      },
...
      "PortalAccessIn": {
        "properties": {
          "email": {
            "maxLength": 254,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "PortalAccessIn",
        "type": "object"
      },
...
      "QueryIn": {
        "properties": {
          "message": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "QueryIn",
        "type": "object"
      },
      "QueryOut": {
        "properties": {
          "access_event_id": {
            "format": "uuid",
            "title": "Access Event Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/QueryStatus"
          }
        },
        "required": [
          "id",
          "patient_id",
          "access_event_id",
          "message",
          "status",
          "created_at"
        ],
        "title": "QueryOut",
        "type": "object"
      },
      "QueryStatus": {
        "enum": [
          "open",
          "answered",
          "closed"
        ],
        "title": "QueryStatus",
        "type": "string"
      },
      "QueryUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/QueryStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "QueryUpdate",
        "type": "object"
      },
...
      "ResourceType": {
        "enum": [
          "demographics",
          "vitals",
          "allergies",
          "notes",
          "prescriptions",
          "lab",
          "billing",
          "documents",
          "consent"
        ],
        "title": "ResourceType",
        "type": "string"
      },
...
      "Role": {
        "enum": [
          "platform_admin",
          "clinic_admin",
          "reception",
          "nurse",
          "doctor",
          "lab_tech",
          "patient"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "Sex": {
        "enum": [
          "female",
          "male",
          "other",
          "unknown"
        ],
        "title": "Sex",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "HealthSaathi API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/access-events/{access_event_id}/query": {
      "post": {
        "operationId": "raise_query_api_v1_access_events__access_event_id__query_post",
        "parameters": [
          {
            "in": "path",
            "name": "access_event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Access Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Raise Query",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/access-queries": {
      "get": {
        "operationId": "list_queries_api_v1_access_queries_get",
        "parameters": [
          {
            "in": "query",
            "name": "open_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Open Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueryOut"
                  },
                  "title": "Response List Queries Api V1 Access Queries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Queries",
        "tags": [
          "portal"
        ]
      }
    },
    "/api/v1/access-queries/{query_id}": {
      "patch": {
        "operationId": "update_query_api_v1_access_queries__query_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "query_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Query Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Query",
        "tags": [
          "portal"
        ]
      }
    },
...
    "/api/v1/me/patient": {
      "get": {
        "operationId": "my_patient_api_v1_me_patient_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Patient",
        "tags": [
          "portal"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/access-log": {
      "get": {
        "operationId": "access_log_api_v1_patients__patient_id__access_log_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessLogPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Access Log",
        "tags": [
          "portal"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/export": {
      "get": {
        "operationId": "export_api_v1_patients__patient_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export",
        "tags": [
          "portal"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/portal-access": {
      "post": {
        "operationId": "portal_access_api_v1_patients__patient_id__portal_access_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalAccessIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Portal Access",
        "tags": [
          "portal"
        ]
      }
    },
...
  }
}